fix branch naming
