+++ asyncgit/src/sync/remotes/mod.rs
@@ -11,10 +11,9 @@ use crate::{
     },
 };
 use crossbeam_channel::Sender;
-use git2::{Direction, FetchOptions, Repository};
+use git2::{FetchOptions, Repository};
 use push::remote_callbacks;
 use scopetime::scope_time;
-use std::collections::HashMap;
 
 /// origin
 pub const DEFAULT_REMOTE_NAME: &str = "origin";
@@ -29,64 +28,6 @@ pub fn get_remotes(repo_path: &str) -> Result<Vec<String>> {
         remotes.iter().flatten().map(String::from).collect();
 
     Ok(remotes)
-}
-
-///
-pub fn get_remotes_branches(
-    repo_path: &str,
-    basic_credential: Option<BasicAuthCredential>,
-) -> Result<HashMap<String, Vec<String>>> {
-    scope_time!("get_remotes_branches");
-
-    let repo = utils::repo(repo_path)?;
-
-    let remotes = repo.remotes()?;
-
-    let mut res = HashMap::with_capacity(remotes.len());
-    for remote in &remotes {
-        if let Some(remote) = remote {
-            let branches = get_remote_branches(
-                &repo,
-                remote,
-                basic_credential.clone(),
-            )?;
-            res.insert(String::from(remote), branches);
-        }
-    }
-
-    Ok(res)
-}
-
-fn get_remote_branches(
-    repo: &Repository,
-    remote: &str,
-    basic_credential: Option<BasicAuthCredential>,
-) -> Result<Vec<String>> {
-    scope_time!("get_remote_branches");
-
-    let mut remote = repo.find_remote(remote)?;
-
-    remote.connect_auth(
-        Direction::Fetch,
-        Some(remote_callbacks(None, basic_credential)),
-        None,
-    )?;
-
-    let list = remote.list()?;
-
-    let res = list
-        .iter()
-        .filter_map(|entry| {
-            let name = entry.name();
-            if name.starts_with("refs/heads/") {
-                Some(String::from(name))
-            } else {
-                None
-            }
-        })
-        .collect();
-
-    Ok(res)
 }
 
 /// tries to find origin or the only remote that is defined if any
@@ -157,15 +98,7 @@ pub(crate) fn fetch_origin(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::sync::{
-        branch::get_branches_info,
-        create_branch,
-        remotes::push::push,
-        tests::{
-            debug_cmd_print, repo_clone, repo_init_bare,
-            write_commit_file,
-        },
-    };
+    use crate::sync::tests::debug_cmd_print;
     use tempfile::TempDir;
 
     #[test]
@@ -291,48 +224,5 @@ mod tests {
         );
         assert_eq!(res.is_err(), true);
         assert!(matches!(res, Err(Error::NoDefaultRemoteFound)));
-    }
-
-    #[test]
-    fn test_remote_branches() {
-        let (r1_dir, _repo) = repo_init_bare().unwrap();
-
-        let (clone1_dir, clone1) =
-            repo_clone(r1_dir.path().to_str().unwrap()).unwrap();
-
-        let clone1_dir = clone1_dir.path().to_str().unwrap();
-
-        // clone1
-
-        write_commit_file(&clone1, "test.txt", "test", "commit1");
-
-        push(clone1_dir, "origin", "master", false, None, None)
-            .unwrap();
-
-        create_branch(clone1_dir, "foo").unwrap();
-
-        write_commit_file(&clone1, "test.txt", "test2", "commit2");
-
-        push(clone1_dir, "origin", "foo", false, None, None).unwrap();
-
-        // clone2
-
-        let (clone2_dir, clone2) =
-            repo_clone(r1_dir.path().to_str().unwrap()).unwrap();
-
-        let clone2_dir = clone2_dir.path().to_str().unwrap();
-
-        let local_branches =
-            get_branches_info(clone2_dir, true).unwrap();
-
-        assert_eq!(local_branches.len(), 1);
-
-        let branches =
-            get_remote_branches(&clone2, "origin", None).unwrap();
-
-        assert_eq!(
-            &branches,
-            &["refs/heads/foo", "refs/heads/master",]
-        );
     }
 }
